Fix some bugs, improve worker to increase mask processing speed
